Calculator.java

+++ Calculator.java
@@ -1,6 +1,6 @@
 public class Calculator {
         public static void main(String[] args) {
-            Double number1, number2, result;
+            int number1, number2, result;
             number1=20;
             number2=10;
                     result = number1 + number2;
